chore: remove the old markdown templates
The markdown templates were weird to navigate since they just dumped
the text in the issue body. This makes the process confusing for users
and does not achieve what we were aiming for, since (in most of the
cases) the users would just delete everything and write the issue on
their own.