feat: add yaml forms for issue templates
Github supports forms for its issue templates and those can be written
in yaml, which is a bit easier to navigate. This commit adds forms for
bug reports and feature requests so that we can better guide the user to
provide all the necessary information we need from the start.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>