feat: Implement build-debian action
This action builds a debian package at a specified path and uploads the
artifacts generated. It is a wrapper around jtdor/build-deb-action,
which itself uses dpkg.

It needs:
- the source dir: where the debian/ folder is located
- docker-image: where the package will be built, and as such the
release to target
- token: This one is optional. If provided, it is used to grant git 
access to private repos, which is necessary to fetch Go private packages

This action has been tested and is in use in a private repository of the
Canonical organization.