Avoid using ${{ inputs.tools-directory }} in bash scripts
When running on windows, the backslashes in the path are ignored by bash
and so the script fails. This happens when cd'ing into the tools
directory.

The solution is to not use cd, and instead rely on the working-directory
field. To avoid excessive boilerplate, I moved all the work that needs
to be done in the tooling directory to the same tooling-version step.