Don't require formatting for commit titles, these cause more problems than they solve.
Adding the type/scope to the commit title is a bad idea for the following reasons:

Encoding data in the title indicates a problem in the tooling. If you need to clearly tag
commits with information, then PR tags should probably be used. These are much more easily
filtered in a UI.

Categories, while familiar to people, are generally a bad way of labelling data.
You can fit the majority of cases into categories but there will always be cases where
no category fits (which means either using no category, a general category or adding
many infrequently used categories). Data also often falls into multiple categories.
This can be handled by listing multiple categories for the data (i.e. treating as tags)
or having to decide a threshold when a category is met. Due to these reasons picking
categories is time consuming task without always a clear solution.

Having a code change, title and type/scope means you now have three things that need to be
consistent. The risk of them not matching is increased over just having two.

A good analogy for this is code comments. The general modern consensus on comments is they
should be used sparingly, and only in cases where the intention of the code can't be easily
determined. If you have a comment that just literally states what the code does it is useless
(this is the same for a PR title where the words match the type/scope). If you have to use
a comment it is an indication that you should consider how your code is written. Rewriting
the code often makes the comment redundant (the same for PR titles - if you need to add the
type/scope then perhaps the title is not appropriately descriptive for the change).

With all the above issues there are likely to be many errors setting the type/scope which
quickly undermines confidence in them. I have experienced this in many projects that use
them.