fix(gh-actions): Use the full command path to go install a tool
Some tools are in subpath of other projects so were not able to install
them properly as per the change in the commit fcbcac1.

So, keep the logic of that commit but rely on unix tools to compute the
full tool path to install, while use the version as go mod told us.