gh-actions: Add action to install gotestfmt and our wrapper script
The output of go test is notoriously hard to read, especially when a lot
of tests are executed and a lot of output is produced. gotestfmt can
hide the output of successful tests and put the output of failed tests
in an expandable group.

This commit adds a custom GitHub Action which installs gotestfmt and a
wrapper script which improves the output and copies it to log files,
which can then be uploaded as test artifacts.