gotestfmt: Add comment to first line of run keyword
In the output of the GitHub Actions run, the commands run for each step
are collapsed, with only the first line being shown unless it's
expanded. It's not helpful if that line is "set -eu" for all our
multi-command steps, so we add a comment to the fist line instead which
describes what the commands are doing.