gh-actions/build-debian: Support running a workflow script at build-source phase
When building some packages we may need to prepare the build environment
before preparing the source package, for examples in some cases we may
need to enable git access (to compute the package version or similar) or
we may need to install tools that are used during debian source building
that are not provided in the archive (such as rust crates, hello
cargo-vendor-filterer!!).

So instead of polluting this generic action with package-specific
requirements, let's just make it support running a script that can
do more advanced configurations in the docker instance that is used
only during the source-preparation phase.