Minimal checks for GH token permissions
It'd be good if we could check more permissions,
but by default a token with `permissions: {}` can still read
a lot of things like existing issues and pull requests, as well
as existing releases and other things. So to start with, I'm only
proposing those two checks, as they are more local to the PR, easy to
spot and less disturbing in case they find issues.

Consider on the other hand checking permission to create an issue. That
requires attempting to create issues, which may be annoying.