Simplifies the commit/push step
We now control the checkout, so we can trust persist-credentials is
true, thus no need to pass the token explicitely on push.
Also we refactor the REF_NAME env var to take into account the
possibility of not being running on a PR context, when github.head_ref
would be null but github.ref is a fully qualified ref name
(refs/heads/SOMETHING...).