Update golangci-lint-action to v8
Since v7, golangci-lint-action supports golangci-lint v2, and doesn't
support v1 anymore. So this is a prerequisite to be able to migrate
our projects to golangci-lint v2, but it also requires us to do so.

golangci-lint v2 doesn't support the "--go" option to specify the
targeted go version. It uses the toolchain version from the go.mod
instead. The "--go" option was the only place where we used the output
of the "go-version" step, so we remove that too.

UDENG-7801