Install govulncheck with the correct Go toolchain version
Unless the GOTOOLCHAIN is specified, `go install` will use the Go
version specified in the go.mod in the current working directory to
build the specified package (govulncheck in this case).

We specify the Go version that should be used for tools like govulncheck
in the tools/go.mod. So we need to point GOTOOLCHAIN to that to actually
use that Go version when building govulncheck.