Set "-eu" directly via the shell command
We can avoid starting each multi-line command with `set -eu` if we just
start the shell with those options.

In fact, the command that's run by the GitHub runner when using just
"bash" as the shell parameter did already use `-e`. The full command run
in that case is:

    bash --noprofile --norc -eo pipefail {0}

Source: https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-syntax#jobsjob_idstepsshell

We just need to add the `-u` option to that, so we get:

    bash --noprofile --norc -euo pipefail {0}