Fix missing export of SOURCE_PATHS in format-source.sh invocation
The SOURCE_PATHS variable was previously assigned in one line and then
./format-source.sh was invoked on the next.

This defined SOURCE_PATHS only as a shell variable, which is visible
within the current shell but not automatically inherited by child
processes. As a result, format-source.sh ran without the variable set,
causing it to ignore the intended source paths.

This change fixes the issue by passing SOURCE_PATHS as part of the
environment for the format-source.sh command.